Move constants into bot
