+++ bot/cogs/emotes.py
@@ -4,12 +4,11 @@ import discord
 from discord.ext.commands import Bot, Cog, Context, command
 from discord.utils import get
 
+from bot import constants
+
 from .emote_utils.emotes import Emote
-from .emote_utils.exceptions import (
-    EmoteNotFoundException,
-    InvalidCommandException,
-)
-from .utils import constants
+from .emote_utils.exceptions import (EmoteNotFoundException,
+                                     InvalidCommandException)
 
 
 class Emotes(Cog):
