+++ bot/cogs/games.py
@@ -4,9 +4,8 @@ from collections import defaultdict
 from discord import Color, Embed, Guild, Member, Message, TextChannel
 from discord.ext.commands import Bot, Cog, Context, command, errors
 
+from bot import constants
 from bot.assets.words import word_list
-
-from .utils import constants
 
 
 class Games(Cog):
